# Refactor: Update schema naming, optimize data extraction based on data contracts, and fix API handlers