--- a/api/dbqueries/language_queries.go
+++ b/api/dbqueries/language_queries.go
@@ -13,21 +13,6 @@ import (
 	"golang.org/x/text/language"
 )
 
-// GetAvailableLanguages fetches all available languages from the database.
-func GetAvailableLanguages() ([]string, error) {
-	return database.GetAvailableLanguages()
-}
-
-// GetLanguageDataTypes fetches available data types for a specific language.
-func GetLanguageDataTypes(lang string) ([]string, error) {
-	return database.GetLanguageDataTypes(lang)
-}
-
-// GetLanguageVersions fetches version information for a specific language.
-func GetLanguageVersions(lang string) (map[string]string, error) {
-	return database.GetLanguageVersions(lang)
-}
-
 // GetLanguageTableData fetches data for a specific language table.
 func GetLanguageTableData(lang, dataType string) (map[string]interface{}, error) {
 	// Construct table name with the new format: ENLanguageDataNounsScribe
@@ -40,7 +25,7 @@ func GetLanguageTableData(lang, dataType string) (map[string]interface{}, error)
 	)
 
 	// Validate table name format and existence
-	if !isValidScribeTableName(tableName) {
+	if !database.IsValidTableName(tableName) {
 		return nil, fmt.Errorf("invalid table name format: %s", tableName)
 	}
 
@@ -69,51 +54,4 @@ func GetLanguageTableData(lang, dataType string) (map[string]interface{}, error)
 		"schema": schema,
 		"data":   data,
 	}, nil
-}
-
-// isValidScribeTableName validates the table name follows the expected pattern.
-func isValidScribeTableName(tableName string) bool {
-	// Expected pattern: ENLanguageDataNounsScribe, FRLanguageDataVerbsScribe
-	if len(tableName) < 10 {
-		return false
-	}
-
-	// Check if it starts with 2-letter language code
-	if len(tableName) < 2 || !isAlpha(tableName[:2]) {
-		return false
-	}
-
-	// Check if it contains LanguageData
-	if !strings.Contains(tableName, "LanguageData") {
-		return false
-	}
-
-	// Check if it ends with Scribe
-	if !strings.HasSuffix(tableName, "Scribe") {
-		return false
-	}
-
-	// Only allow alphanumeric characters
-	for _, char := range tableName {
-		if !isAlphaNumeric(char) {
-			return false
-		}
-	}
-
-	return true
-}
-
-// isAlpha checks if string contains only alphabetic characters.
-func isAlpha(s string) bool {
-	for _, char := range s {
-		if !((char >= 'A' && char <= 'Z') || (char >= 'a' && char <= 'z')) {
-			return false
-		}
-	}
-	return true
-}
-
-// isAlphaNumeric checks if character is alphanumeric.
-func isAlphaNumeric(char rune) bool {
-	return (char >= 'A' && char <= 'Z') || (char >= 'a' && char <= 'z') || (char >= '0' && char <= '9')
 }