--- a/api/dbqueries/language_queries.go
+++ b/api/dbqueries/language_queries.go
@@ -15,7 +15,7 @@ import (
 
 // GetLanguageTableData fetches data for a specific language table.
 func GetLanguageTableData(lang, dataType string) (map[string]any, error) {
-	// Construct table name with the new format: ENLanguageDataNounsScribe
+	// Construct table name with the new format: ENLanguageDataNounsScribe.
 
 	caser := cases.Title(language.English)
 
@@ -24,12 +24,12 @@ func GetLanguageTableData(lang, dataType string) (map[string]any, error) {
 		caser.String(dataType),
 	)
 
-	// Validate table name format and existence
+	// Validate table name format and existence.
 	if !database.IsValidTableName(tableName) {
 		return nil, fmt.Errorf("invalid table name format: %s", tableName)
 	}
 
-	// Check if table exists
+	// Check if table exists.
 	exists, err := database.TableExists(tableName)
 	if err != nil {
 		return nil, fmt.Errorf("error checking table existence for %s: %w", tableName, err)
@@ -38,13 +38,13 @@ func GetLanguageTableData(lang, dataType string) (map[string]any, error) {
 		return nil, fmt.Errorf("table %s does not exist", tableName)
 	}
 
-	// Get table schema
+	// Get table schema.
 	schema, err := database.GetTableSchema(tableName)
 	if err != nil {
 		return nil, fmt.Errorf("error fetching schema for %s: %w", tableName, err)
 	}
 
-	// Get data from table
+	// Get data from table.
 	data, err := database.GetTableData(tableName)
 	if err != nil {
 		return nil, fmt.Errorf("error fetching data for %s: %w", tableName, err)