--- a/api/server.go
+++ b/api/server.go
@@ -23,11 +23,6 @@ func HandleRequests() {
 		log.Fatalf("Failed to initialize database: %v", err)
 	}
 
-	// Create language_data_versions table if it doesn't exist.
-	if err := database.CreateLanguageDataVersionsTable(); err != nil {
-		log.Printf("Warning: Failed to create language_data_versions table: %v", err)
-	}
-
 	// Set Gin mode based on environment.
 	switch viper.GetString("GIN_MODE") {
 	case "release":
@@ -96,10 +91,11 @@ func startServer(r *gin.Engine) {
 	validators.InitLanguageValidator(availableLanguages)
 
 	log.Printf("👀 Listening on port %s", hostPort)
-	log.Printf("🚀 API endpoints available:")
-	log.Printf("  ✅ GET /api/v1/data/:language_iso - Versioned API")
-	log.Printf("  ✅ GET /api/v1/data-version/:language_iso - Version check API")
-	log.Printf("  ✅ GET /api/v1/languages - List available languages")
+	log.Println("🚀 API Endpoints:")
+	log.Println("  ✅ GET /api/v1/languages                		- List available languages")
+	log.Println("  ✅ GET /api/v1/contracts[?lang_iso=xx]      	- Get contracts (optional language filter)")
+	log.Println("  ✅ GET /api/v1/data/:lang_iso       		- Get full language data with schema")
+	log.Println("  ✅ GET /api/v1/data-version/:lang_iso 		- Get version info for a language")
 	log.Printf("📊 Available languages: %v", availableLanguages)
 
 	log.Fatal(r.Run(hostPort))