--- a/api/server.go
+++ b/api/server.go
@@ -36,17 +36,17 @@ func HandleRequests() {
 	// Create Gin router with default middleware (logger and recovery).
 	r := gin.Default()
 
-	// proxies trust security warning
+	// Proxies trust security warning.
 	trustedProxies := []string{} // Default empty (trusts nothing)
 	if os.Getenv("ENV") == "prod" {
-		// Trust RFC 1918 private network ranges commonly used by cloud providers and load balancers
+		// Trust RFC 1918 private network ranges commonly used by cloud providers and load balancers.
 		trustedProxies = []string{
 			"10.0.0.0/8", // Class A private range (10.0.0.0 - 10.255.255.255)
-			// Used by: AWS VPC, Google Cloud, Kubernetes clusters, most cloud providers
+			// Note: Used by AWS VPC, Google Cloud, Kubernetes clusters, most cloud providers.
 			"172.16.0.0/12", // Class B private range (172.16.0.0 - 172.31.255.255)
-			// Used by: Docker default bridge networks, some enterprise networks
+			// Note: Used by Docker default bridge networks, some enterprise networks.
 			"192.168.0.0/16", // Class C private range (192.168.0.0 - 192.168.255.255)
-			// Used by: Home routers, small office networks, some internal services
+			// Note: Used by Home routers, small office networks, some internal services.
 		}
 		log.Printf("🔒 Production mode: trusting proxy networks")
 	} else {