… config-example.yaml
… contracts/de.json
… contracts/en.json
… contracts/es.json
… contracts/fr.json
… contracts/it.json
… contracts/pt.json
… contracts/ru.json
… contracts/sv.json
… database/language.go
… database/table.go
… database/versions.go
… internal/constants/strings.go
… main.go
… models/models.go
… update_data.sh