<!---
Thank you for your pull request! 🚀
-->

### Contributor checklist

<!-- Please replace the empty checkboxes [] below with checked ones [x] accordingly. -->

- [x] This pull request is on a [separate branch](https://docs.github.com/en/get-started/quickstart/github-flow) and not the main branch
- [x] I have ran the `./pre-commit` executable as well as `make lint` and have fixed all reported issues

---

### Description

<!--
Describe briefly what your pull request proposes to change. Especially if you have more than one commit, it is helpful to give a summary of what your contribution is trying to solve.

Also, please describe shortly how you tested that your change actually works.
-->
#### Summary

This PR makes key updates to the language data pipeline and API handlers:

1. Table naming convention
    - Changed from ENLanguageData_nouns → ENLanguageDataNounsScribe
    - Matches the naming scheme requested by @andrewtavis  for consistency across all supported languages and data types.
    
2. Data extraction improvements
    - Focused script only on converting filtered data into SQLite instead of entire dumps.
    - Pipeline steps:
        - Download lexemes dump file
        - Extract data for 8 supported languages
        - Filter data based on contracts in Scribe-Data repo
        - Convert filtered JSON -> SQLite
        - Perform DB migrations
     
    - Performance gain: Reduced extraction from ~7.55s to <400ms for English tables.
    ## 📸 Before vs After
    
    | Before | After |
    |--------|-------|
    | <img width="450" alt="After" src="https://github.com/user-attachments/assets/c7a2463e-02a6-4ec1-a5b8-b9bce009186f" /> | <img width="450" alt="Before" src="https://github.com/user-attachments/assets/881afc18-0c50-495c-9593-e87b412a2b00" />|


3. API handler fixes
    - Updated queries to use new table names & field naming conventions.
    - Ensures endpoints now fetch data correctly from the updated database schema.
    
#### Testing
To test this branch locally:

1. **Pull the THIS branch**

2. **Start MySQL**
   Open your terminal and run:

   ```bash
   mysql
   ```

3. **Update database schema**
   Run the update script in another terminal:

   ```bash
   make update
   ```

4. **Run the server**

   ```bash
   make run
   ```

5. **Test the endpoints**
   Once the server is running, you can hit the following endpoints:

   * `GET /api/v1/data/:language_iso` – Versioned API
   * `GET /api/v1/data-version/:language_iso` – Version check API
   * `GET /api/v1/languages` – List available languages


### Related issue

<!--- Scribe-Server prefers that pull requests be related to already open issues. -->
<!--- If applicable, please link to the issue by replacing ISSUE_NUMBER with the appropriate number below. -->
<!--- Feel free to delete this section if this does not apply. -->

- closes #18 as it uses the contracts provided by `Scribe-Data` to perform JSON data filtration.