@axif0 2025-08-22
in the [IsLanguageSupported](https://github.com/scribe-org/Scribe-Server/blob/main/api/validators/language_validator.go) function, can we use like - 
```
func IsLanguageSupported(lang string, availableLanguages []string) bool {
	return slices.Contains(availableLanguages, lang)
}
```

I think it is more cleaner and shorter.