@DeleMike 2025-08-25
Hi @axif0, can you review now?

You can test the process by running `make update-data` then `make dev`