@andrewtavis 2025-08-25
The expected response from the endpoint would be the JSON of the contract. As of now as I understand it we have the contracts saved on the server, and we'd want a way to deliver those with the data. That way if we need to update the contract and the data that it provides, the steps would be:

1. Update the contracts on Scribe-Server
2. Run a new process to get the data
3. Filter the data by the new contracts
4. Create the MariaDB tables
5. Serve both the contract and the data

An example is like what I was saying above about verb conjugations where in the contract they have numbered indexes that then are loaded into the conjugation UI in order. Say that we have two verb tenses - present and past - and we want to add in future tense. We'd need add in the needed entry for future tense to the contract, get new data and filter it, prepare the data, and then the clients would need both the new data and the contract that was used to filter it.