@DeleMike 2025-08-28
Hmm, I don't really understand why...

Please ensure you add `contractsDir: "./contracts"` to the `config.yaml`
And also, can I see your logs? 
do you see something like this: `2025/08/28 04:16:00 📊 Available languages: [de en es fr it pt ru sv]`

<details>

<summary>See my full logs:</summary>

2025/08/28 04:15:59 ✅ Database connection established
2025/08/28 04:15:59 📖 API Documentation available at: /swagger/index.html
2025/08/28 04:15:59 Serving files from: ./
2025/08/28 04:16:00 👀 Listening on port :8080
2025/08/28 04:16:00 🚀 API Endpoints:
2025/08/28 04:16:00   ✅ GET /api/v1/languages                          - List available languages
2025/08/28 04:16:00   ✅ GET /api/v1/contracts[?lang_iso=xx]            - Get contracts (optional language filter)
2025/08/28 04:16:00   ✅ GET /api/v1/data/:lang_iso                     - Get full language data with schema
2025/08/28 04:16:00   ✅ GET /api/v1/data-version/:lang_iso             - Get version info for a language
2025/08/28 04:16:00 📊 Available languages: [de en es fr it pt ru sv]
[GIN] 2025/08/28 - 04:18:12 | 200 |   11.223184ms |             ::1 | GET      "/api/v1/contracts?lang=en"
[GIN] 2025/08/28 - 04:18:42 | 200 |     4.30634ms |             ::1 | GET      "/api/v1/contracts"
[GIN] 2025/08/28 - 04:30:14 | 200 |    4.450882ms |             ::1 | GET      "/api/v1/contracts?lang=fr"
[GIN] 2025/08/28 - 04:30:20 | 200 |     1.36483ms |             ::1 | GET      "/api/v1/contracts?lang=en"

</details>

If you do not see any languages, it should be because we don't support the old tables, i.e `EN_LanguageData_nouns` but rather we support `ENLanguageDataNounsScribe`. 

```sql
SELECT DISTINCT SUBSTRING(TABLE_NAME, 1, 2) as language_code
FROM information_schema.TABLES 
WHERE TABLE_SCHEMA = ? 
AND TABLE_NAME LIKE '%LanguageData%Scribe'
ORDER BY language_code
```

So maybe you should run `make update-data`, then try `make dev` again, or you can simply just do `make migrate` and `make dev`

---
looking forward to your response!