@axif0 2025-08-29
> `make dev` again,

Now it works :sweat_smile: