@axif0 2025-08-29
I'm not sure, why I got this error :thinking: 
```
You trusted all proxies, this is NOT safe. We recommend you to set a value.
Please check https://pkg.go.dev/github.com/gin-gonic/gin#readme-don-t-trust-all-proxies for details.
```

<img width="915" height="205" alt="image" src="https://github.com/user-attachments/assets/923fc3b3-ad7b-4e75-bd51-cc0cca3887e8" />