@DeleMike 2025-08-30
NEW COMMANDS UPDATE:

# Development (default behaviour)
`make run`          # Same as `make run-dev`
`make run-dev`      # Explicit development
`make dev`          # Hot reload development

# Production
`make run-prod`     # Production mode
`make dev-prod`     # Hot reload production mode