@DeleMike 2025-08-30
## 🔒 Configure trusted proxies for security

This new commit addresses Gin's proxy security warning by explicitly configuring trusted proxy networks.

- **Dev**: Trust no proxies (direct connections)
- **Prod**: Trust RFC 1918 private ranges (`10.0.0.0/8`, `172.16.0.0/12`, `192.168.0.0/16`)
- **Added**: `make run-dev` and `make run-prod` targets

**RFC 1918** private ranges are IP address blocks reserved for internal networks that can't be accessed from the public internet.

10.0.0.0/8 = 10.0.0.1 to 10.255.255.255 (cloud providers, Kubernetes, for example which Toolforge uses)
172.16.0.0/12 = 172.16.0.1 to 172.31.255.255 (Docker networks)
192.168.0.0/16 = 192.168.0.1 to 192.168.255.255 (home/office routers)

Why trust them: Only legitimate infrastructure (load balancers, proxies) can use these IPs - external attackers can't spoof them because internet routers block these addresses from public networks.

**Learn more:** 
- [Gin proxy docs](https://pkg.go.dev/github.com/gin-gonic/gin#Engine.SetTrustedProxies)
- [RFC 1918 private networks](https://tools.ietf.org/html/rfc1918)